+++ src/main/java/fun/rubicon/util/SafeMessage.java
@@ -2,6 +2,7 @@ package fun.rubicon.util;
 
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.MessageEmbed;
 import net.dv8tion.jda.core.entities.TextChannel;
 
 import java.util.concurrent.TimeUnit;
@@ -33,8 +34,23 @@ public class SafeMessage {
     }
 
     public static void sendMessage(TextChannel textChannel, String message, int deleteTime) {
-        if (hasPermissions(textChannel))
+        if (hasPermissions(textChannel) && hasDeletePermission(textChannel))
             textChannel.sendMessage(message).queue(msg -> msg.delete().queueAfter(deleteTime, TimeUnit.SECONDS));
+        else if(hasPermissions(textChannel))
+            textChannel.sendMessage(message).queue();
+
+    }
+
+    public static void sendMessage(TextChannel textChannel, MessageEmbed build) {
+        if (hasPermissions(textChannel))
+            textChannel.sendMessage(build).queue();
+    }
+
+    public static void sendMessage(TextChannel textChannel, MessageEmbed build, int deleteTime) {
+        if (hasPermissions(textChannel) && hasDeletePermission(textChannel))
+            textChannel.sendMessage(build).queue(msg -> msg.delete().queueAfter(deleteTime, TimeUnit.SECONDS));
+        else if(hasPermissions(textChannel))
+            textChannel.sendMessage(build).queue();
     }
 
     public static Message sendMessageBlocking(TextChannel textChannel, String message) {
@@ -43,10 +59,19 @@ public class SafeMessage {
         return null;
     }
 
+
     private static boolean hasPermissions(TextChannel channel) {
         if (channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_READ) && channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_WRITE))
             return true;
         channel.getGuild().getOwner().getUser().openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(EmbedUtil.error("Permission Error!", "The bot need the `MESSAGE_READ` and `MESSAGE_WRITE` permissions in the ``" + channel.getName() + "` channel to run without errors.").build()));
         return false;
     }
+    private static boolean hasDeletePermission(TextChannel channel){
+        if(channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_MANAGE))
+            return true;
+        channel.getGuild().getOwner().getUser().openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(EmbedUtil.error("Permission Error!", "The bot need the `MESSAGE_READ` and `MESSAGE_MANAGE` permissions in the ``" + channel.getName() + "` channel to run without errors.").build()));
+        return false;
+    }
+
+
 }
